Makefiles updated in src and test folders
