+++ Irrigation/test/Commons/LoggerTest.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <gmock/gmock.h>
 
 
 class LoggerTest : public ::testing::Test {
