+++ Irrigation/test/Logic/Zones_test.h
@@ -1,6 +1,6 @@
 #pragma once
+#include <gmock/gmock.h>
 
-#include "gtest/gtest.h"
 #include "Logic/Zones.h"
