+++ Irrigation/test/Utils/Parse_test.cpp
@@ -1,4 +1,5 @@
 #include "common.h"
+#include <gmock/gmock.h>
 #include "Utils/Parse.h"
