+++ Irrigation/test/Views/CommandLineView/Commands_test.h
@@ -1,5 +1,5 @@
 #pragma once
-
+#include <gmock/gmock.h>
 #include <memory>
 
 #include "Model/IrrigationDocument.h"
