+++ Irrigation/test/Views/CommandLineView/Makefile
@@ -1,28 +1,29 @@
 
-TARGET_NAME     = CommandLineView
+TARGET_NAME     = $(shell basename $(CURDIR))
 
 ########################################################
 ########################################################
 
 PROJECT_ROOT    = ../../..
 SRC_EXT         = cpp
-SOURCE_DIR      = $(PROJECT_ROOT)/test/$(TARGET_NAME)
-OBJECT_DIR      = $(PROJECT_ROOT)/build/test/$(TARGET_NAME)
+SOURCE_DIR      = $(PROJECT_ROOT)/test/Views/$(TARGET_NAME)
+OBJECT_DIR      = $(PROJECT_ROOT)/build/test/Views/$(TARGET_NAME)
 INCLUDE_DIR     = $(PROJECT_ROOT)/include
 TARGET_DIR      = $(PROJECT_ROOT)/build
 
 TARGET          = $(TARGET_DIR)/$(TARGET_NAME)_test.exe
-SOURCES         = $(shell find $(SOURCE_DIR) -type f -name '*.$(SRC_EXT)')
+SOURCES         = $(shell find $(SOURCE_DIR) -maxdepth 1 -type f -name '*.$(SRC_EXT)')
 OBJECTS         = $(patsubst $(SOURCE_DIR)/%,$(OBJECT_DIR)/%,$(SOURCES:.$(SRC_EXT)=.o))
+DEPENDS         = $(patsubst $(SOURCE_DIR)/%,$(OBJECT_DIR)/%,$(SOURCES:.$(SRC_EXT)=.d))
 
 ########################################################
 ########################################################
 
-INCLUDE_PATH    = -I$(INCLUDE_DIR) -I$(INCLUDE_DIR)/$(TARGET_NAME)
-LIBDIRS         = -L$(TARGET_DIR)
-LIBS            = -lgmock -l$(TARGET_NAME) 
+INCLUDE_PATH    = -I$(INCLUDE_DIR)
+LIBDIRS         = -L$(TARGET_DIR) -L/usr/local/lib
+LIBS            = -lgmock -l$(TARGET_NAME) -lModel -lLogic -lHardware -lCommons -lCommand -lDocumentView -lWebServerView -lWebServer -lteng -lmicrohttpd -lUtils
 
-CXXFLAGS        = -std=gnu++11 -Wall -fmessage-length=0
+CXXFLAGS        = -std=gnu++11 -Wall -fmessage-length=0 -MMD -MP 
 CPPFLAGS        = -isystem $(PROJECT_ROOT)/lib/gmock_gtest $(INCLUDE_PATH) 
 
 # DEBUG flags
@@ -32,25 +33,25 @@ CPPFLAGS        = -isystem $(PROJECT_ROOT)/lib/gmock_gtest $(INCLUDE_PATH)
 CXXFLAGS        += -O3
 
 
-all: test
+all:  $(TARGET)
+	@echo 'Executing test: $(TARGET)'
+	@$(TARGET)
+	@echo 'Finished executing test: $(TARGET)'
+	@echo ' ' 
 
-$(OBJECT_DIR)/%.o: $(SOURCE_DIR)/%.$(SRC_EXT)
+-include $(DEPENDS)
+
+$(OBJECT_DIR)/%.o: $(SOURCE_DIR)/%.$(SRC_EXT) 
 	@mkdir -p $(OBJECT_DIR)
-	@echo 'Building file: $<'
-	@echo 'Invoking: Cygwin C++ Compiler'
 	$(CXX) $(CXXFLAGS) $(CPPFLAGS) -c -o $@ $<
-	@echo 'Finished building: $<'
-	@echo ' '
 
-$(TARGET): $(OBJECTS)
-	@echo 'Building target: $@'
-	@echo 'Invoking: Cygwin C++ Linker'
+$(TARGET):	$(OBJECTS)
+	@mkdir -p $(TARGET_DIR)
 	$(CXX) $(LIBDIRS) -o $(TARGET) $(OBJECTS) $(LIBS)
-	@echo 'Finished building target: $@'
-	@echo ' '
 
-clean:	
+clean:
 	rm -f $(OBJECTS)
+	rm -f $(DEPENDS)
 	rm -f $(TARGET)
 
 print:
@@ -60,11 +61,5 @@ print:
 	@echo TARGET_DIR: $(TARGET_DIR)
 	@echo SOURCES: $(SOURCES)
 	@echo OBJECTS: $(OBJECTS)
+	@echo DEPENDS: $(DEPENDS)
 	@echo TARGET: $(TARGET)
-
-test: $(TARGET)
-	@echo 'Executing test: $(TARGET)'
-	@$(TARGET)
-	@echo 'Finished executing test: $(TARGET)'
-	@echo ' ' 
-
