+++ Irrigation/test/Views/CommandLineView/test_main.cpp
@@ -1,0 +1,7 @@
+#include <gmock/gmock.h>
+
+
+int main(int argc, char **argv) {
+	::testing::InitGoogleMock(&argc, argv);
+	return RUN_ALL_TESTS();
+}
